gardenbook_chat_api/gardenbook_chat/plants-mcp/pyproject.toml
[project]
name = "garden-mcp"
version = "0.1.0"
description = "Garden Book plant management server using MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.0",
    "pymongo>=4.5.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0", 
]

[tool.pytest]
testpaths = ["tests"]
